Rename reactive-messaging extension to smallrye-reactive-messaging
