+++ bom/pom.xml
@@ -83,7 +83,7 @@
 
          <dependency>
             <groupId>org.jboss.shamrock</groupId>
-            <artifactId>shamrock-reactive-messaging-deployment</artifactId>
+            <artifactId>shamrock-smallrye-reactive-messaging-deployment</artifactId>
             <version>${project.version}</version>
             <scope>provided</scope>
          </dependency>
