--- extensions/reactive-messaging/deployment/pom.xml
@@ -1,55 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <parent>
-    <artifactId>shamrock-reactive-messaging</artifactId>
-    <groupId>org.jboss.shamrock</groupId>
-    <version>1.0.0.Alpha1-SNAPSHOT</version>
-    <relativePath>../</relativePath>
-  </parent>
-
-  <modelVersion>4.0.0</modelVersion>
-
-  <artifactId>shamrock-reactive-messaging-deployment</artifactId>
-  <name>Shamrock - Reactive Messaging - Deployment</name>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.jboss.shamrock</groupId>
-      <artifactId>shamrock-core-deployment</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.jboss.shamrock</groupId>
-      <artifactId>shamrock-arc-deployment</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.jboss.shamrock</groupId>
-      <artifactId>shamrock-reactive-messaging-runtime</artifactId>
-    </dependency>
-    
-    
-    <dependency>
-      <groupId>org.jboss.shamrock</groupId>
-      <artifactId>shamrock-junit5-internal</artifactId>
-    </dependency>
-    
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <artifactId>maven-compiler-plugin</artifactId>
-        <configuration>
-          <annotationProcessorPaths>
-            <path>
-              <groupId>org.jboss.shamrock</groupId>
-              <artifactId>shamrock-extension-processor</artifactId>
-              <version>${project.version}</version>
-            </path>
-          </annotationProcessorPaths>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
-</project>
