--- extensions/reactive-messaging/deployment/src/test/java/org/jboss/shamrock/reactivemessaging/SimpleBean.java
@@ -1,42 +1,0 @@
-package org.jboss.shamrock.reactivemessaging;
-
-import java.util.List;
-import java.util.concurrent.CopyOnWriteArrayList;
-
-import javax.annotation.PostConstruct;
-import javax.enterprise.context.ApplicationScoped;
-
-import org.eclipse.microprofile.reactive.messaging.Incoming;
-import org.eclipse.microprofile.reactive.messaging.Outgoing;
-import org.eclipse.microprofile.reactive.streams.operators.PublisherBuilder;
-import org.eclipse.microprofile.reactive.streams.operators.ReactiveStreams;
-
-@ApplicationScoped
-public class SimpleBean {
-    
-    static final List<String> RESULT = new CopyOnWriteArrayList<>();
-    
-    @Outgoing("source")
-    public PublisherBuilder<String> source() {
-      return ReactiveStreams.of("hello", "with", "SmallRye", "reactive", "message");
-    }
-
-    @Incoming("source")
-    @Outgoing("processed-a")
-    public String toUpperCase(String payload) {
-      return payload.toUpperCase();
-    }
-
-    @Incoming("processed-a")
-    @Outgoing("processed-b")
-    public PublisherBuilder<String> filter(PublisherBuilder<String> input) {
-      return input.filter(item -> item.length() > 4);
-    }
-
-    @Incoming("processed-b")
-    public void sink(String word) {
-      System.out.println(">> " + word);
-      RESULT.add(word);
-    }
-
-}
