--- extensions/reactive-messaging/runtime/pom.xml
@@ -1,57 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-   <parent>
-      <artifactId>shamrock-reactive-messaging</artifactId>
-      <groupId>org.jboss.shamrock</groupId>
-      <version>1.0.0.Alpha1-SNAPSHOT</version>
-      <relativePath>../</relativePath>
-   </parent>
-   <modelVersion>4.0.0</modelVersion>
-
-   <artifactId>shamrock-reactive-messaging-runtime</artifactId>
-   <name>Shamrock - Reactive Messaging - Runtime</name>
-
-   <dependencies>
-      <dependency>
-         <groupId>org.jboss.shamrock</groupId>
-         <artifactId>shamrock-arc-runtime</artifactId>
-      </dependency>
-      <dependency>
-         <groupId>org.jboss.shamrock</groupId>
-         <artifactId>shamrock-reactive-streams-operators-runtime</artifactId>
-      </dependency>
-      <dependency>
-         <groupId>io.smallrye.reactive</groupId>
-         <artifactId>smallrye-reactive-messaging-provider</artifactId>
-         <exclusions>
-            <exclusion>
-               <groupId>com.fasterxml.jackson.core</groupId>
-               <artifactId>jackson-core</artifactId>
-            </exclusion>
-         </exclusions>
-      </dependency>
-   </dependencies>
-
-   <build>
-      <plugins>
-         <plugin>
-            <artifactId>maven-dependency-plugin</artifactId>
-         </plugin>
-         <plugin>
-           <artifactId>maven-compiler-plugin</artifactId>
-           <configuration>
-             <annotationProcessorPaths>
-               <path>
-                 <groupId>org.jboss.shamrock</groupId>
-                 <artifactId>shamrock-extension-processor</artifactId>
-                 <version>${project.version}</version>
-               </path>
-             </annotationProcessorPaths>
-           </configuration>
-         </plugin>
-      </plugins>
-   </build>
-
-</project>
