--- extensions/reactive-messaging/runtime/src/main/java/org/jboss/shamrock/reactivemessaging/runtime/ReactiveMessagingLifecycle.java
@@ -1,31 +1,0 @@
-package org.jboss.shamrock.reactivemessaging.runtime;
-
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutionException;
-
-import javax.enterprise.context.Dependent;
-import javax.enterprise.event.Observes;
-import javax.inject.Inject;
-
-import org.jboss.shamrock.runtime.StartupEvent;
-
-import io.smallrye.reactive.messaging.extension.MediatorManager;
-
-@Dependent
-public class ReactiveMessagingLifecycle {
-
-    @Inject
-    MediatorManager mediatorManager;
-
-    void onApplicationStart(@Observes StartupEvent event) {
-        CompletableFuture<Void> future = mediatorManager.initializeAndRun();
-        try {
-            future.get();
-        } catch (ExecutionException e) {
-            throw new RuntimeException(e.getCause());
-        } catch (InterruptedException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-}
