+++ extensions/smallrye-reactive-messaging/deployment/src/test/java/org/jboss/shamrock/smallrye/reactivemessaging/StreamConsumer.java
@@ -1,0 +1,27 @@
+package org.jboss.shamrock.smallrye.reactivemessaging;
+
+import java.util.List;
+
+import javax.enterprise.context.ApplicationScoped;
+import javax.inject.Inject;
+
+import org.eclipse.microprofile.reactive.messaging.Message;
+
+import io.reactivex.Flowable;
+import io.smallrye.reactive.messaging.annotations.Stream;
+
+@ApplicationScoped
+public class StreamConsumer {
+
+    @Inject
+    @Stream("source")
+    private Flowable<Message<String>> sourceStream;
+
+    public List<String> consume() {
+        return Flowable.fromPublisher(sourceStream)
+                .map(Message::getPayload)
+                .toList()
+                .blockingGet();
+    }
+
+}
