+++ extensions/smallrye-reactive-messaging/pom.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+   <parent>
+      <artifactId>shamrock-build-parent</artifactId>
+      <groupId>org.jboss.shamrock</groupId>
+      <version>1.0.0.Alpha1-SNAPSHOT</version>
+      <relativePath>../../build-parent/pom.xml</relativePath>
+   </parent>
+
+   <modelVersion>4.0.0</modelVersion>
+   <artifactId>shamrock-smallrye-reactive-messaging</artifactId>
+   <name>Shamrock - SmallRye Reactive Messaging</name>
+   <packaging>pom</packaging>
+
+   <modules>
+      <module>deployment</module>
+      <module>runtime</module>
+   </modules>
+</project>
